+++ CHANGELOG.md
@@ -9,6 +9,11 @@ This Changelog tracks changes to this project. The notes below include a summary
 - `fixed` for any bug fixes.
 - `security` in case of vulnerabilities.
 
+## Version `0.27.3` - 05 Oct 2022
+
+- `changed` schemas bump for MIBI support
+- `added` facets for MIBI files
+
 ## Version `0.27.2` - 01 Oct 2022
 
 - `added` PACT User role